third_party/blink/web_tests/http/tests/xmlhttprequest/encode-request-url-2-expected.txt

Test how URLs are encoded.

abs-path: PASS
rel-path: PASS
absolute: PASS
Frame, abs-path: PASS
Frame, rel-path: PASS
 
